chore: Revert Flashbar fixes (#787)
