+++ src/app-layout/notifications/index.tsx
@@ -10,21 +10,16 @@ interface NotificationsProps {
   children?: React.ReactNode;
   labels: AppLayoutProps.Labels | undefined;
   topOffset: number | undefined;
-  disableContentPaddings?: boolean;
 }
 interface NotificationWrapperProps extends NotificationsProps {
   sticky: boolean | undefined;
 }
 
 export const Notifications = React.forwardRef(
-  ({ sticky, disableContentPaddings, ...props }: NotificationWrapperProps, ref: React.Ref<HTMLDivElement>) => {
+  ({ sticky, ...props }: NotificationWrapperProps, ref: React.Ref<HTMLDivElement>) => {
     return sticky ? (
       <div ref={ref} className={styles['notifications-sticky']} style={{ top: props.topOffset }}>
-        <div
-          role="region"
-          className={clsx(props.testUtilsClassName, disableContentPaddings && styles['no-content-paddings'])}
-          aria-label={props.labels?.notifications}
-        >
+        <div role="region" className={props.testUtilsClassName} aria-label={props.labels?.notifications}>
           {props.children}
         </div>
       </div>
@@ -32,11 +27,7 @@ export const Notifications = React.forwardRef(
       <div
         role="region"
         ref={ref}
-        className={clsx(
-          props.testUtilsClassName,
-          styles.notifications,
-          disableContentPaddings && styles['no-content-paddings']
-        )}
+        className={clsx(props.testUtilsClassName, styles.notifications)}
         aria-label={props.labels?.notifications}
       >
         {props.children}
