+++ src/app-layout/visual-refresh/notifications.tsx
@@ -25,11 +25,6 @@ export default function Notifications() {
     return null;
   }
 
-  /*
-  The notificationsElement ref is assigned to an inner div to prevent internal bottom margin from affecting the
-  calculated height, which is used for sticky elements below.
-   */
-
   return (
     <div
       role="region"
@@ -44,8 +39,9 @@ export default function Notifications() {
         testutilStyles.notifications,
         'awsui-context-content-header'
       )}
+      ref={notificationsElement}
     >
-      <div ref={notificationsElement}>{notifications}</div>
+      {notifications}
     </div>
   );
 }
