+++ src/flashbar/collapsible-flashbar.tsx
@@ -20,7 +20,6 @@ import { sendToggleMetric } from './internal/analytics';
 import { useFlashbar } from './common';
 import { throttle } from '../internal/utils/throttle';
 import { scrollElementIntoView } from '../internal/utils/scrollable-containers';
-import { findUpUntil } from '../internal/utils/dom';
 
 export { FlashbarProps };
 
@@ -113,7 +112,7 @@ export default function CollapsibleFlashbar({ items, ...restProps }: FlashbarPro
           const windowHeight = window.innerHeight;
           // Take the parent region into account if using the App Layout, because it might have additional margins.
           // Otherwise we use the Flashbar component for this calculation.
-          const outerElement = findUpUntil(flashbar, element => element.getAttribute('role') === 'region') || flashbar;
+          const outerElement = flashbar.parentElement?.parentElement || flashbar;
           const applySpacing =
             isFlashbarStackExpanded && Math.ceil(outerElement.getBoundingClientRect().bottom) >= windowHeight;
           if (!applySpacing) {
@@ -286,10 +285,7 @@ export default function CollapsibleFlashbar({ items, ...restProps }: FlashbarPro
         styles.flashbar,
         styles[`breakpoint-${breakpoint}`],
         styles.stack,
-        isCollapsible && styles.collapsible,
-        items.length === 2 && styles['short-list'],
         isFlashbarStackExpanded && styles.expanded,
-        isVisualRefresh && styles['visual-refresh'],
         getVisualContextClassname('flashbar')
       )}
       ref={mergedRef}
@@ -302,8 +298,7 @@ export default function CollapsibleFlashbar({ items, ...restProps }: FlashbarPro
               styles['notification-bar'],
               isVisualRefresh && styles['visual-refresh'],
               isFlashbarStackExpanded ? styles.expanded : styles.collapsed,
-              transitioning && styles['animation-running'],
-              items.length === 2 && styles['short-list']
+              transitioning && styles['animation-running']
             )}
             onClick={toggleCollapseExpand}
             ref={notificationBarRef}
